requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-money==2.0.1
django-phonenumber-field==6.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py-moneyed==1.2
pycodestyle==2.7.0
pycparser==2.20
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-money==2.0.1
django-phonenumber-field==6.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py-moneyed==1.2
pycodestyle==2.7.0
pycparser==2.20
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-money==2.0.1
django-phonenumber-field==6.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py-moneyed==1.2
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.14.1
pytz==2021.1
rave-python==1.2.15
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-money==2.0.1
django-phonenumber-field==6.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py-moneyed==1.2
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.14.1
pytz==2021.1
rave-python==1.2.15
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
